Bs-34 | removed all fhirterminody api and omod reference
